Test Harness for 3270...
git-svn-id: https://jintn3270.svn.sourceforge.net/svnroot/jintn3270/trunk@71 d6bd4e80-8031-4e1d-bd39-93c407d940fc

+++ src/com/sf/jintn3270/actions/BackspaceTerminalAction.java
@@ -13,6 +13,6 @@ public class BackspaceTerminalAction extends TerminalAction {
 	
 	public void doAction(TerminalModel model) throws Exception {
 		model.eraseChar();
-		model.getClient().sendCommand(model.getClient().EC);
+		model.getClient().sendCommand((byte)model.getClient().EC);
 	}
 }
